requirements.txt
absl_py==2.1.0
gym==0.26.2
gymnasium==0.29.1
numpy==2.1.1
PySC2==3.0.0
setuptools==69.5.1
stable_baselines3==2.3.2
